+++ ext/log-processor/src/main/java/core/log/service/ActionLogForwarder.java
@@ -2,6 +2,7 @@ package core.log.service;
 
 import core.framework.kafka.MessagePublisher;
 import core.framework.log.message.ActionLogMessage;
+import core.log.LogForwardConfig;
 
 import java.util.HashSet;
 import java.util.List;
@@ -15,20 +16,23 @@ public class ActionLogForwarder {
     private final String topic;
     private final Set<String> apps;
     private final Set<String> results;
+    private final Set<String> ignoreActions;
     private final Set<String> ignoreErrorCodes;
 
-    public ActionLogForwarder(MessagePublisher<ActionLogMessage> publisher, String topic, List<String> apps, List<String> results, List<String> ignoreErrorCodes) {
+    public ActionLogForwarder(MessagePublisher<ActionLogMessage> publisher, LogForwardConfig.Forward forward) {
         this.publisher = publisher;
-        this.topic = topic;
-        this.apps = new HashSet<>(apps);
-        this.results = new HashSet<>(results);
-        this.ignoreErrorCodes = new HashSet<>(ignoreErrorCodes);
+        this.topic = forward.topic;
+        this.apps = new HashSet<>(forward.apps);
+        this.results = new HashSet<>(forward.results);
+        this.ignoreActions = new HashSet<>(forward.ignoreActions);
+        this.ignoreErrorCodes = new HashSet<>(forward.ignoreErrorCodes);
     }
 
     public void forward(List<ActionLogMessage> messages) {
         for (ActionLogMessage message : messages) {
             if (apps.contains(message.app)
                 && (results.isEmpty() || results.contains(message.result))
+                && !ignoreActions.contains(message.action)
                 && !ignoreErrorCodes.contains(message.errorCode)) {
                 publisher.publish(topic, null, message);
             }